+++ src/cli_wizard/constants.py
@@ -3,7 +3,7 @@
 
 """Constants for CLI Wizard."""
 
-from importlib.metadata import version, PackageNotFoundError
+from importlib.metadata import PackageNotFoundError, version
 
 try:
     __version__ = version("cli-wizard")